requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
python-dateutil==2.9.0.post0
httpx==0.28.1
beautifulsoup4==4.12.3
feedparser==6.0.11
